smart-subscription/pyproject.toml
[project]
name = "smart-subscription"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bcrypt==4.1.2",
    "flask==3.0.3",
    "flask-cors==4.0.1",
    "mongoengine==0.28.2",
    "pyjwt==2.8.0",
    "pymongo==3.13.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "razorpay==1.4.1",
    "setuptools<70",
]
